Add default realm to example krb5.conf
This example caused the test playbook to fail with an error saying that the Kerberos config did not specify a default realm.

Once I added that extra line to the configmap, it worked great. Just thought I would contribute that back as a PR for anyone else that finds this in the future.

Thanks!